experiment with generating docs for package:flutter
