+++ benchmark/benchmark.md
@@ -2,10 +2,14 @@
 
 ---
 
+### 2023-11-01
+
+package:flutter: 22.9s (4,495 files, 79.3MB)
+
 ### 2023-10-31
 
-Dart SDK: 3.9s (1278 files, 13.3MB)
+Dart SDK: 3.9s (1,278 files, 13.3MB)
 
 ### 2023-10-25
 
-Dart SDK: 5.6s (1277 files, 199.6MB)
+Dart SDK: 5.6s (1,277 files, 199.6MB)
